The de-lighting is the memory ceiling, and a flaky tile server is survivable
Black Bear Pass at 8192 samples was OOM-killed, exit 137, and so - it turns out -
was the point-cloud fetch that "died mysteriously" hours earlier. imagery.py
conditions the base colour over the whole array at once, so the working set scales
with base_tex_px squared and not with the terrain's sample count: the stage sat at
2.5 GB through compositing and passed 11 GB when it reached the de-lighting. It is
the photograph, not the ground, that decides whether a map builds.

So base_tex_px drops to 4096 on Black Bear Pass (2 m per texel over 8192 m, the
density Bingham Canyon and Mt St Helens already ship) with a comment saying why,
and three float64 DEM copies become float32 - 537 MB each at 8192 squared, bought
for nothing, since float32 holds a millimetre at terrestrial magnitudes and
ndimage allocates its own copy alongside whatever it is handed. AGENTS.md records
the ceiling, the exit-137 signature, and what tiling the de-lighting would take.

Wallace Creek is re-specced to the 16,384 m square at 1 m the user chose: Soda
Lake, the Goodwin Education Center, Elkhorn Road, Wallace Creek and its
interpretive trail, and the Temblor crest, centred 35.2215 N 107.8310 W so the
tightest landmark keeps 790 m and the offset channel 2.6 km. Reaching KCL and
Fellows would have meant the next power-of-two rung, 32.8 km: four times the
ground for 5.3 GB, 7.5 billion returns and twelve hours of reading.

And USGS's 3DEP ImageServer returns 504 over the Carrizo Plain for requests it
serves elsewhere, then 502s a 128 px request minutes later - it is intermittently
sick, not size-bound. A tile the server refuses now splits into quadrants and is
mosaicked back into one file, so the cache layout and tile names do not change;
and a tile that still fails is deferred to a second pass at the end instead of
aborting the run. One bad tile in sixty-four used to lose the other sixty-three.

The Black Bear road extract was also refetched: the tracked copy was still the old
4096 m bbox with 42 ways, so Telluride would have shipped with no streets. 373 now.

+++ examples/gis_maps/maplib/imagery.py
@@ -42,7 +42,7 @@ def fit_sun(
     from scipy import ndimage
 
     lum = colour_u8.astype("float32").mean(axis=-1)
-    smooth = ndimage.gaussian_filter(dem.astype("float64"), 2.0)
+    smooth = ndimage.gaussian_filter(dem.astype("float32"), 2.0)
     gy, gx = np.gradient(smooth, res)
     slope = np.arctan(np.hypot(gx, gy))
     aspect = np.arctan2(-gx, gy)
@@ -180,7 +180,7 @@ def delight(
         res_r = res * dem.shape[0] / colour_u8.shape[0]
     else:
         dem_r, res_r = dem, res
-    smooth = ndimage.gaussian_filter(dem_r.astype("float64"), 1.0)
+    smooth = ndimage.gaussian_filter(dem_r.astype("float32"), 1.0)
     shade = hm.hillshade(smooth, res_r, azimuth_deg, altitude_deg)
     visibility = cast_shadows(smooth, res_r, azimuth_deg, altitude_deg)
     # Sky light is what the shadows see, and a crater floor sees less sky than a plain:
@@ -1466,7 +1466,7 @@ def aspect_flatfield(
 
     from scipy import ndimage
 
-    smooth = ndimage.gaussian_filter(dem.astype("float64"), 2.0)
+    smooth = ndimage.gaussian_filter(dem.astype("float32"), 2.0)
     gy, gx = np.gradient(smooth, res)
     slope = np.degrees(np.arctan(np.hypot(gx, gy)))
     aspect = np.arctan2(gx, -gy) % (2 * np.pi)  # 0 = north-facing, clockwise
